backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
argon2-cffi==23.1.0
cryptography==41.0.7
# pysqlcipher3==1.2.0  # Optional: for production with SQLCipher
# For development, we use standard sqlite3
python-dotenv==1.0.0
PyJWT==2.8.0
email-validator==2.1.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
bandit==1.7.5
ollama==0.6.1
requests==2.31.0